requirements.txt
Pillow
piexif
requests
geopy
openai
vllm
numpy
opencv-python
python-dateutil
tqdm
jsonschema
transformers>=4.57.0
torch
accelerate
qwen-vl-utils
scipy
timm
einops
ftfy
torchvision

decord>=0.6.0
beautifulsoup4
lxml
fastapi
uvicorn
python-multipart
